+++ forui/lib/src/foundation/item/item_content.dart
@@ -14,7 +14,7 @@ class ItemContent extends StatelessWidget {
   final FItemContentStyle style;
   final FWidgetStateMap<FDividerStyle>? dividerStyle;
   final FItemDivider dividerType;
-  final EdgeInsetsGeometry padding;
+  final EdgeInsetsGeometry margin;
   final Set<WidgetState> states;
   final Widget? prefix;
   final Widget title;
@@ -26,7 +26,7 @@ class ItemContent extends StatelessWidget {
     required this.style,
     required this.dividerStyle,
     required this.dividerType,
-    required this.padding,
+    required this.margin,
     required this.states,
     required this.title,
     required this.prefix,
@@ -42,8 +42,8 @@ class ItemContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => ItemContentLayout(
-    padding: padding,
-    margin: style.padding,
+    margin: margin,
+    padding: style.padding,
     dividerStyle: dividerStyle?.resolve(states),
     dividerType: dividerType,
     children: [
@@ -113,7 +113,7 @@ class ItemContent extends StatelessWidget {
       ..add(DiagnosticsProperty('style', style))
       ..add(DiagnosticsProperty('dividerStyle', dividerStyle))
       ..add(DiagnosticsProperty('dividerType', dividerType))
-      ..add(DiagnosticsProperty('padding', padding))
+      ..add(DiagnosticsProperty('margin', margin))
       ..add(IterableProperty('states', states));
   }
 }